Restructure AccumulateAllComponents.jsx and index.jsx
